requirements.txt
# Elasticsearch Agent Builder Hackathon - Dependencies

# Core dependencies (REQUIRED)
elasticsearch>=8.11.0
python-dotenv>=1.0.0
urllib3>=1.26.0

# Web framework for demo UI
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# Data visualization for metrics
matplotlib>=3.8.0
seaborn>=0.13.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Utilities
colorama>=0.4.6
